Fix nextjs boilerplate
